Fix compiling without sqlite and gcc
